refactor index.d.ts file for Firebase ML
